Now only the available  tickets are shown
Changes to the db: I have change the names of the columns on the db. Now are more describable.
Run the following query to rebuild the table.

CREATE TABLE `student` (
  `TICKETID` int(11) NOT NULL,
  `EVENTNAME` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
  `LOCATION` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
  `PRICE` varchar(50) COLLATE utf8_unicode_ci DEFAULT NULL,
   `PASSWORD` varchar(50) COLLATE utf8_unicode_ci DEFAULT NULL,
  `YEARLEVEL` int(11) DEFAULT NULL,
  PRIMARY KEY (`TICKETID`)
) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci

Now the index.jsp actually shows only the available tickets. (The button still dosen't work but I will fix it asap. )
Only the tickets which the "PASSWORD" column in the db set to 0 are available. If the "PASsWORD" column is not 0 the ticket is considered "reserved" and wont be displayed.